+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.9.1] - 2019-03-10
+### Fixed
+* Allow user to ctrl-c to terminate piping-server docker container
+### Changed
+* Return "Content-Length" and "Content-Type" when accessing index, /version and /help
+
 ## [0.9.0] - 2019-03-08
 ### Changed
 * Make logs for sender consistent
@@ -143,7 +149,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 * Docker automated build on Docker Hub
 * Support HTTPS
 
-[Unreleased]: https://github.com/nwtgck/piping-server/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/nwtgck/piping-server/compare/v0.9.1...HEAD
+[0.9.1]: https://github.com/nwtgck/piping-seraver/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/nwtgck/piping-seraver/compare/v0.8.10...v0.9.0
 [0.8.10]: https://github.com/nwtgck/piping-seraver/compare/v0.8.9...v0.8.10
 [0.8.9]: https://github.com/nwtgck/piping-seraver/compare/v0.8.8...v0.8.9